# htdocs/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cooking Book API",
        "description": "REST API for Cooking Book application - manage recipes, comments, likes and categories",
        "contact": {
            "email": "admin@cookingbook.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/admin/recipes": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get all recipes (Admin only)",
                "operationId": "9562f6bc43aaf0b7c4577533cd54ac78",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search in title",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Filter by user",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Filter by category",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Sort field",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "title",
                                "created_at",
                                "views_count",
                                "updated_at"
                            ]
                        }
                    },
                    {
                        "name": "sort_order",
                        "in": "query",
                        "description": "Sort order",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of all recipes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "description": "Recipe object",
                                                "type": "object"
                                            }
                                        },
                                        "metadata": {
                                            "properties": {
                                                "pagination": {
                                                    "properties": {
                                                        "current_page": {
                                                            "type": "integer"
                                                        },
                                                        "per_page": {
                                                            "type": "integer"
                                                        },
                                                        "total": {
                                                            "type": "integer"
                                                        },
                                                        "last_page": {
                                                            "type": "integer"
                                                        },
                                                        "from": {
                                                            "type": "integer",
                                                            "nullable": true
                                                        },
                                                        "to": {
                                                            "type": "integer",
                                                            "nullable": true
                                                        },
                                                        "has_more": {
                                                            "type": "boolean"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "next_cursor": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "filters": {
                                                    "properties": {
                                                        "search": {
                                                            "type": "string",
                                                            "nullable": true
                                                        },
                                                        "user_id": {
                                                            "type": "integer",
                                                            "nullable": true
                                                        },
                                                        "category_id": {
                                                            "type": "integer",
                                                            "nullable": true
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "sortings": {
                                                    "properties": {
                                                        "sort_by": {
                                                            "type": "string",
                                                            "nullable": true
                                                        },
                                                        "sort_order": {
                                                            "type": "string",
                                                            "nullable": true
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/recipes/{id}": {
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Force delete recipe (Admin only)",
                "operationId": "03013167e88fc384b975f077b57d4fd2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recipe deleted"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Recipe not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}